Create codebook script.

Create the complete codebook latex script.

Usage

codebook(
  varInfo,
  missings,
  struc,
  scaleInfo,
  register = NULL,
  dat,
  Kennwertedatensatz,
  chapters,
  deckblatt = "",
  intro = "",
  literatur = "",
  abkuerzverz = "",
  hintmod = "",
  lastpage = ""
)

Arguments

varInfo

data.frame or list of data.frames containing the variable information, imported via getVarInfo.

missings

data.frame or list of data.frames containing the missing information, imported via getMissings.

struc

data.frame containing the structure of the codebook, imported via getStructure.

scaleInfo

data.frame containing the information on scales, imported via getScaleInfo.

register

data.frame containing the information on the register, imported via getRegister. If NULL, now register is created. If there are registers for some data sets but not all, the missing registers are simply omitted.

dat

data.frame or list of data.frames containing the data sets, imported via import_spss.

Kennwertedatensatz

data.frame or list of data.frame containing the descriptive statistics, imported via calculateDescriptives.

chapters

data.frame or list of data.frames containing the chapter information, imported via getChapters. Determines the order of chapters in the codebook.

deckblatt

Character vector with the cover page, created via makeCover.

intro

Character vector, introduction.

literatur

Character vector with the literature information, created via makeLit.

abkuerzverz

Character vector with the abbreviations, created via makeAbbrList.

hintmod

Character vector with the information on the background model, created via makeBGM.

lastpage

Character vektor, last page.

Value

Codebook latex script.

Examples

#tbd